requirements.txt
openai>=1.12.0
groq>=0.4.0
flask>=3.0.0
python-dotenv>=1.0.0
colorama>=0.4.6
